Hyper-V: Implement nova rescue
The root disk image is moved to a separate disk slot while the
rescue image will take it's place. If the instance requires it,
a temporary config drive is created as well.

Unrescuing the instance will move the root disk image back in
place, removing temporary images.

Implements: blueprint hyper-v-rescue

Change-Id: I6059ae35a77d675f54b98b2b43b5762e1d24365b