Adds Hyper-V OVS ViF driver
This branch adds a new ViF driver to the Nova Hyper-V compute driver.
It's responsible for creating the OVS port and the related association
with the Hyper-V virtual switch port. Aside from the openvswitch binaries
it has no other dependencies. Flow rules are added by an external agent
such as OpenDaylight or by a Neutron OVS agent.

Change-Id: I4cee52f65968fa503f2c8a2c9a206489719fb15a
Implements: blueprint hyper-v-ovs-vif