Fix serial console handler cleanup
The named pipe handler stop method checks if the 'stopped' event
was set before attempting to perform cleanup. The issue is if one
worker fails, it will set this event, preventing the stop method
from performing the cleanup.

In this case, the VM console log files will remain open, unable to
be deleted when the VM is destroyed.

This check can be safely removed.

The serial console handler is now stopped in case some workers
fail to start, avoiding the above mentioned situation.

Change-Id: I6b431cc530b87defeb0b10e1886f030a452b4b06