Fix event handler NotFound error
When an instance is destroyed, it is stopped first. The manager
will get the event and consider it an unexpected power off, so
it will try to call the stop API as well.

Meanwhile, the instance might be deleted, in which case the above
operation will fail with an "NotFound" error.

This can be avoided by catching this in the power off method.

Change-Id: Ibd173bfb6cefa7cd585eee2ff48b583841b1c05c