Fix QoS issues caused by unsupported OS versions
Disk QoS is not available on versions prior to WS 2012 R2.

Because of this and the fact that the driver attempts to set IOPS
limits even if those are not specified, spawning instances currently
fails when using previous HyperV versions.

This patch fixes the issues by skipping setting QoS specs when
this feature is not supported, logging a warning.

Closes-Bug: #1463317

Change-Id: I086f3a034ffcc56a09a5ecf3e9a142cc86e67c41
(cherry picked from commit b97666aa6d897280a6a31d5ff6fd8d9eab3b7a4d)