Let the eventhandler manage serial console workers
At the moment, if an instance is unexpectedly power off or paused,
although the serial console worker threads are stopped, there is no
cleanup performed. For example, the serial proxy will still be
accepting connections.

This patch lets the event handler manage serial console workers
based on the recieved events.

Change-Id: I154e5796698dc6a57defb588c5d47f4f095fe7de