Fixes "Hyper-V destroy vm fails on Windows Server 2008R2"
Delete vm fails on Windows Server 2008R2 with the following error:

[Error 145] The directory is not empty: 'C:\\OpenStack\\Instances\\instance-00000003'

This is happening because on Windows 2008R2 it takes a while (1-2 sec. )
for Hyper-V to delete the instance specific files that it stores in the
instance folder.

This patch fixes this bug by adding a retry in the delete method if
it gets that specific error: WindowsError 145

Closes Bug: 1467409
Change-Id: I6d907abde256575cd58b368578dd4b940603c619
(cherry picked from commit c972154be23e12bd097eacc6cae8ccc82b9d9c2b)