Hyper-V: Fixes method retrieving free SCSI controller slot on V1
The method retrieving a free SCSI controller slot gets all the
related disk resources, checking their address using the
AddressOnParent attribute.

The issue is that this WMI object attribute is not available on the
V1 virtualization namespace, for which reason this method will raise
an AttributeError in case there disks connected to the SCSI
controller.

This patch fixes the issue by properly retrieving the disk address.

Closes-Bug: #1467916

Change-Id: I7f88e37ac6d13cd8a15b17939ec8dbdca2c988e1