Hyper-V: Lock snapshot operation using instance uuid
The instance snapshot operation is not synchronized withing the
compute manager.

Attempting to destroy an instance while it's being snapshoted fails
when using Hyper-V. This can lead to leaked instances.

In order to avoid this, this patch synchronizes the instance
snapshot operation within the Hyper-V driver.

The alternative would be canceling pending WMI jobs when instance
termination is requested, which we consider introducing as a later
effort.

Change-Id: I1ad3f795925503e59bf51ee3148871a88e9eecf0
Partial-Bug: #1461970