Fixes Hyper-V boot from volume fails when using ephemeral disk
Instance boot from volume fails in Hyper-V if the flavor has ephemeral
disk.

This happens because Hyper-V tries to attach both the ephemeral disk and
the boot volume to the same slot on the IDE controller.

This fixes the bug by making sure that the ephemeral disk will always
attach to the second slot.

Change-Id: Ia4679d008c4061053a56ad6d087165597b317361
Closes-Bug: #1479129