Fix unrescue operation in case of already mounted configdrive
If the rescue operation fails, the instance is attempted to be
unrescued.

If the original configdrive was not detached during the
rescue operation, the unrescue operation will fail as it
will attempt to attach an already attached disk.

This patch fixes the issue by checking if the configdrive is
already mounted before attempting to reattach it.

Closes-bug: #1496794

Change-Id: I64884dad80d9594663c97c2afc9bdac3b26a2310
(cherry picked from commit 4f29c9ff11fe505f0b31c12f443766b1cdd5c28d)