Add retry decorator to methods adding/modifying resources
Adding virtual resources was consistently failing on our CIs,
especially when attaching the configdrive.

This patch attempts to fix this by adding the same decorator
that was used for the modify_virt_resource method, which had
avoided similar issues.

This is applied to the _modify_virtual_system method as well
for the same reason.

Change-Id: I1d3b0deb0682857b5cc688a25876868ecc26bcf0