Fixes 'VMMS Crash when querying MSVM_Processor'
Due to a bug in vmms, getting Msvm_Processor for the numa
node associators resulted in a vmms crash.
As an alternative to using associators we have to manually get
the related Msvm_Processor classes.
Msvm_HostedDependency is the association class between
Msvm_NumaNode and Msvm_Processor. We need to use this class to
relate the two because using associators on Msvm_Processor
will also result in a crash.

Closes-Bug: #1538543

Change-Id: Ie3675b75fe275e03247f065d35b8e3a4d01ac271