hyperv: Drop JSON for supported_instances
The virt drivers pass the value for supported_instances
to the resource tracker as a JSON encoded string. This
gets decoded as soon as it is copied to the compute node
object. The JSON is not used anymore since compute node
object was added so we will remove the encoding for all
virt drivers.

This patch removes that unnecessary encoding from the
hyperv driver.

(cherry picked from commit 45482fa94a351d29bddfd6c36262dcf84c75d600)

blueprint mitaka-objects

Change-Id: I532621c7ed0320c89fb4e0125655f465c69ab6e8