Adds lookup VM as VSSDs
A lot of vmutils and vmutilsv2 operations require
a Msvm_VirtualSystemSettingData. Currently, it is fetched
from a Msvm_ComputerSystem associators, which is an expensive
operation. The needed object can simply be queried directly.
Using associators is a time expensive operation, replacing them
with direct queries improves the performance by decreasing the
time needed to wait to retrieve the associated classes.

Co-Authored-By: Claudiu Belu <cbelu@cloudbasesolutions.com>
Change-Id: If78f413a374920ed43d1148ed91457049c4d4eda