Fixes compatibility with the WMI module
In the case of association classes, there were some differences
between PyMI and the WMI module in the way they were returning
class properties that were references to objects. This patch
fixes the compatibility with the WMI module by expecting that the
class properties that are references to be returned as objects.

Currently, PyMI returns the reference while WMI returns the object.
This patch depends on a change in PyMI which will make it behave
the same way as the WMI module does in case of class properties that
are references.

The WMI module does not handle building queries properly if the
path is passed directly as parameter in the case of association
classses. The query string has to be built manually to maintain
the compatibility with the WMI module.

Change-Id: Idedecd88779f82438c853778f3355289537504e2