Fix retry decorators
Some of our vmutils methods use loopingcall.RetryDecorator,
having a maximum number of retries. The issue with that is that
the implementation of this decorator uses a static counter for
the number of retries performed.

In other words, after the maximum number of retries is performed,
no more retries will be performed even if the method is called
in a different context.

This patch adds a new decorator, fixing this issue.

(cherry-picked from commit 420eb02ab912cf1c103cdda22c0079e133aa88cc)

Use the retry decorator for methods adding/removing virt res

In some cases, the add_virt_resource method fails because of what
seems to be a Hyper-V race condtion.

This change adds a retry decorator for this method, as well as the
one removing virt resources.

(cherry-picked from commit 52d4bbfef2004dbdcf3a42a486beae5158e9fdb7)

NOTE(claudiub): removed error_codes from original commit, as the
HyperVExceptions do not have error codes.

Change-Id: I7645db47f02b145e21fb44a8508753e808ab7212