Fix missing parameter lookup_vm when retrieving VM id
The VM id is currently retrieved from a field of a wmi object
of type Msvm_ComputerSystem. The method lookup_vm is missing
a parameter to return the needed wmi object, and it returns
an object of type Msvm_VirtualSystemSettingData which doesn't
have the field containing the VM id.

Trace: http://paste.openstack.org/show/e1bCNzN2s6QUJvU5jZvb/

Change-Id: Iddf47ff42f9d4526b88cd3bca9c24ddb17761526