hyper-v: Copies back files on failed migration
On cold migration, the contents of the instance folder are
copied to the new host. The original folder cannot be removed
because the VM configuration files cannot be deleted until the VM
is destroyed.

Because of this, when the migration fails to copy the files, it will
try to revert this through folder renaming. Since the original folder
still exists, an exception is raised.

Change-Id: Ia42ed873924999d57336a105bcaa2b856f3a3a9d
Closes-Bug: #1555699
(cherry picked from commit 7fb627c317d93b041b76263740f5729dade05b39)