SMBFS: Ensure the attach/disconnect operations are synchronized
We need to make sure that those operations are synchronized so that
the share is not disconnected while we are trying to attach a disk.

This was already done previously, but the patch adding FC support
has moved the attach method to the BaseVolumeDriver.

This change fixes this issue. Also, the decorator used for
synchronizing the operations based on the share path has been moved
to the SMBFS volume driver, as it's strictly related to it.

Change-Id: Ia840741f4fcde5e5d56020b7ee7068833095fa4c