Properly handle shared storage in case of cold migrations
At the moment, if the destination host is other than the source
host, we attempt to move the instance files without checking if
shared storage is being used.

This change fixes this issue. In case of shared storage, we'll have
the same workflow as when the migration source is the migration
destination as well.

Closes-Bug: #1565895

Change-Id: I711976b42a5de5a28de22e010b48c10d96b68ba5