Fixes Nova compute driver import
A recent change in nova made it so only drivers from the
nova.virt namespace are imported.

Removes tools/tox_install.sh, it was only needed to
install nova, which was needed for unit tests.

Also, the nova.netconf options are not imported explicitly
anymore as those have been moved to nova.conf.

Change-Id: Idaae1a0a06aa0385b11176ea129a54c294f25fab