FibreChannel: Fix missing device
At the moment, if the device path was found, we are not performing
any rescan.

This can cause issues if we retrieve the device path before it was
claimed by the MPIO service (in which case, the device number can
change). If the device number changes, the instances won't be
able to access it anymore. Also, detaching the disk will fail.

Simply performing a rescan before attempting to fetch the volume
mappings fixes this issue.

Closes-Bug: #1597002
Change-Id: Ic72afda7839c6949f3e833061d1143ae9e49fe16
(cherry picked from commit cf25f5c62a678672b93537ce4adac7fd5dcfb677)