Remove unnecessary mock
In one of the unit tests, we unnecessarily mock os.path.join. This
seems to cause the test to hang on Python 3.4, as gettextutils will
hang when attempting to translate one of the messages logged by
the tested method.

For now, we'll just remove this mock as it's blocking the gate.
Still,We may have to understand better why we did not have this
issue before.

Change-Id: I00b0e1d2dd10a392096c35e81d402d5a4b0ef703