Fix live migration with shared storage
At the moment, when live migration is requested, we attempt to
copy the instance dvd disks and log files to the remote location,
other instance files being transparently copied by Hyper-V.

The issue is that we don't check whether shared storage is being
used, in which case this will fail.

This change adds this check.

Closes-Bug: #1565895

Change-Id: Ib646c90c830a1cd0e5401d14c6d400226b034f73