Create planned VMs during LM only when needed
When performing a live migration, a planned VM is created every
time a volume is attached to the instance, while this is required
only in case of passthrough disks.

Skipping creating the planned VM when there are only disk files
attached takes out an unnecessary step, also dropping the Kerberos
constrained delegation requirement for environments where iSCSI/FC
based storage backends are not being used for Cinder.

Change-Id: Ib1aad4b5cf013b2ac62dc797ca6c53a2932f953f