Removes unnecessary config option for RemoteFX
The RemoteFX support patch was recently merged in nova.
Since RemoteFX is already a part of compute-hyperv,
the config option "enable_remotefx" is being registered
in compute-hyperv.

As a consequence of the patch merging in nova, the config
option "enable_remotefx" is now being registered in nova.
Due to this, the nova-compute service fails to start when
using compute-hyperv because it will try to register the
same config option twice.

This patch addresses this issue by removing the config option
registration from compute-hyperv.

Change-Id: I1b22ad13ae9f75a60716190bb46d0495ae28b8c2