HyperV: remove instance snapshot lock
At the moment, the instance snapshot operation is synchronized using
the instance uuid. This was added some time ago, as the instance
destroy operation was failing when an instance snapshot was in
proggress.

This is now causing a deadlock, as a similar lock was recently
introduced in the manager for the shelve operation by this change:
Id36b3b9516d72d28519c18c38d98b646b47d288d

We can safely remove the lock from the HyperV driver as we now stop
pending jobs when destroying instances.

Closes-Bug: #1611321

Change-Id: I1c2ca0d24c195ebaba442bbb7091dcecc0a7e781