Properly handle instances when not placed at configured location
In some situations, we may have to handle instances that are located
at a different path than the configured one. For example, this can
happen in case of failed over instances when multiple CSVs are used
among hosts.

This change will let the driver locate the instance files in this
situation.

Depends-On: I56c1394fdd3e93598e1b41d354c072ae06264201
Depends-On: Ia7623a689854f937d70323cf38818d7534f3ccd9

Change-Id: Ibf78880413dbed6acdb0bca76010893621e944e2
Closes-Bug: #1621334