Properly handle shared storage during live migration
After a live migration is performed, we always try to delete the
instance files.

If the migration fails, the destination node is not cleaned up as
the HyperVLiveMigrateData object is not used at the moment and the
nova compute manager won't know whether a cleanup is needed or not.

Also, in order to support using CSVs for storing instance files, we
need to explicitly specify whether image files will be migrated or not
when requesting os-win to migrate the VM.

This change addresses those issues.

Depends-On: I48dc6da29534bf0c477c742311ac15438d8b63ad

Change-Id: Icf2f258ef8900b5970ff153a7b981ee53e58dd2d
Partial-Bug: #1565895