refactors cold migration / resize
Simplifies the cold resize / migration process by making the
destination node pull the VM files from the source instead of
vice-versa. Doing so, the logic is simplified greatly, the
destination node only has to copy the files, regardless if it
is on shared storage or not.

Solves the migration issue when the compute nodes have different
instances_path config option set, as the destination node will
copy the files to its configured instances_path folder.

Co-Authored-By: Claudiu Belu <cbelu@cloudbasesolutions.com>

Change-Id: Ia44850675e8cb8eb21017e1c4321802574dfe6a0