Add automatic host shutdown instance action option
At the moment, the instances will be automatically saved whenever
the host is shut off. This adds a disk overhead which may not be
desired.

For this reason, this patch makes it configurable. By turning on
'instance_automatic_shutdown', instances will be turned off instead
of being saved when the host is turned off.

Note that existing VMs will not be affected by this config option.
For this reason, whenever an admin will change it, he has to make
sure that the existing VMs are updated.

Instance disk overhead will be calculated based on the assumption
that all the instances are configured the same way.

Depends-On: I68490a4e257eef88bdd07f3ec14d345a2df79a4a

Change-Id: I46c1497fe1863b09c5f41ddd0cbd6a23e9e4828c