don't move files on migration
In some cases (shared storage, CSVs), moving the instance
files on cold migration / resize might not be desired.

Adds move_disks_on_cold_migration config option (default value True).

If the config option is True, or if the source instance path is not
on a CSV or a share, the instance files are copied to the destination
node. Otherwise, the files are kept on the source node.

Depends-On: Ic153dcfd621a1e6235dd3a75241274ee7d7c868b

Change-Id: Ia093d6e68d033d060afff863c22ef2d97490b0ff