Check failed cluster vm live migrations
At the moment, when migrating clustered instances, we only catch
HyperVVMNotFoundException exceptions, in which case we attempt to
recover the instance. For any other exception, the instance will
hang in a 'migrating' state.

This change checks the instance state, trying to recover if it's
not in an inconsistent state, setting it to 'error' state otherwise.

Depends-On: I6c268355213718d4d66e0e77a89c558cde3213b6

Closes-Bug: #1674438

Change-Id: I952122bc2f0386f68c0ad62d195cfc5d84648876