Fixes timeout waiting for vif plugging callback for instance
Currently, plug_vifs is called when powering on the instance.
The call to wait_for_plug_events will give a timeout waiting
for the plug event when using neutron-ovs-agent.

This patch addresses this issue by moving the call to plug_vifs
from power_on to after the create_instance call, which is prior
to waiting for plug events. This change modifies the behaviour
of power_on only during the instance spawn, the other calls
remaining the same.

Change-Id: I13b7caf84434a4b9b6f2795cddc7401874c46647
Closes-Bug: #1684250