Remove log translations
Starting with the Pike series, OpenStack no longer supports log
translation.
This patch remove log translations then update hacking rules to prevent
future mistakes. It also enables H904, which allows the logging package
to skip creating the formatted log message if the message is not going
to be emitted because of the current log level. [1]

https://docs.openstack.org/oslo.i18n/latest/user/guidelines.html#adding-variables-to-log-messages

Change-Id: I584d10356acea0ede6d219a8ca9e1b001b8d6064