Improve detaching passthrough disks
Retrieving the disk path can be a time consuming operation in case
of passthrough disks. As such disks attachments will be tagged using
the volume id, we'll just use that instead.

Note that Hyper-V does not allow us to attach the same passthrough
disk to multiple instances, which means that we're safe to rely
on this tag.

Depends-On: Ic3775c3e81a5f05b20221ef52e393fdb54d0190c

Change-Id: I99448881d8f9e247e1c34b1cd489c5488eff7146