Validate retrieved disk resources
When attaching passthrough disks, we're passing a Msvm_DiskDrive
object path instead of the actual disk path.

The issue is that even if the according disk path is retrieved, we
may fail to retrieve the above mentioned resource path.

This change adds the missing check. This way, a retry may be
performed if needed. Also, this ensures that we don't get a VM disk
with no actual disk attached (Hyper-V allows it).

Closes-Bug: #1709049

Change-Id: I55abf82f51d4dc739083eec2ea479d2c8115670f