Ensure that vif ports are unplugged during failover migration
There is a race condition during failover migration, that, when
triggered, stops the host from proper cleanup.
If the destination node finishes the failover before the host,
proper cleanup is done, otherwise there will pe an ovs port
remaining on the source host.

This change addresses this issue by removing the race condition
which was previously used to check if the migration was done
properly.
This doesn't completly fix the because, if the host was already updated,
no aciton is taken at the source node.

Change-Id: I50213de235907f80e1d4266dfd6785ea147a7f0c
Partial-Bug: #1705992