Ignore external VMs when handling serial console output
When the driver is initialized, we're attempting to capture serial
console output for instances that were not created by Nova,
although those should be skipped.

The event handler decides whether an instance should be skipped or
not based on its instance notes, which are expected to contain an
uuid, the instance uuid, in case of Nova instances.

We'll do the same when the driver is initialized. Note that there
may be VMs that were not created by us but do have an UUID in their
"Notes" field but this workflow is *good enough* and consistent for
now.

Alternatively, we could fetch from nova a list of instances that are
expected to reside on this node, but this may be out of sync. The
other option would be to change the notes format, adding some specific
prefix, but that would not be backwards compatible.

Change-Id: I156348e39a0db80160a7223d18dd9cfd277e7e5a
Closes-Bug: #1709057