Validate SAN policy when having passthrough disks
One common cause for passthrough disk attachment issues is an
incorrectly configured san policy. Having it set to 'online' will
prevevent disks from being attached to VMs.

This change allows the hyperv driver to check the san policy and
log a warning message when the driver loads and when a passthrough
disk is being attached.

Depends-On: I25033c15b77f494a417c9cd01d194a8bfb3cfe13
Change-Id: Ia123d98286fa99f45f91bfc47a4f0d468fa7d13e
Closes-Bug: #1727257