Fixing signature mismatch for spawn method
Nova was unable to spawn instances on compute nodes because of an
argument not being passed.

More specific, the 'spawn' method of the 'HyperVClusterDriver'
class was called with 9 arguments, when the signature only had 8.
This would raise an error as one of the arguments was receiving
two values instead of one.

The cluster driver public method signatures are now validated against
the base Nova driver.

This change fixes this issue.

Change-Id: I086a5c6a859bc4fa2424e840cb390462d8cdff85