Fix instance disk resize issues
This change provides the following instance disk resize fixes:

- removed ephemeral disks are detached during the resize process,
currently this step is skipped, for which reason the planned vms
cannot be *realized*.

- when resizing from a flavor with no ephemeral storage to a new
one that does have ephemeral storage, no ephemeral disk is currently
added as there is no explicit ephemeral BDM. This is inconsistent
with the Libvirt driver behavior as well as Hyper-V driver behavior
prior to the new BDM format addition. This change ensures that we're
adding an ephemeral disk in such situations.

- If the new flavor does not provide ephemeral storage, all ephemeral
disks are removed.

- We're now validating the new flavor ephemeral storage size before
performing the resize/migration, thus failing early in case of an
invalid flavor, also avoiding error state instances.

- We're no longer erroring out if the new flavor's root disk size
is smaller than the current one and the instance was booted from a
volume.

Closes-Bug: 1724240

Change-Id: I296afc9bd947021fcca9199f6aca95ab51cd7df9