Raise proper exception in case of vif plug errors
The Nova Compute manager expects a VirtualInterfacePlugException
exception to be raised when a vif plug fails.

This particularly affects the service initialization step. If we're
raising an exception other than the above mentioned one, the exception
won't be cought, in which case the Nova Compute service will stop.

This change ensures that we're raising the expected exception if vif
plugging fails.

Change-Id: Ibaa405577d4fa06f27fdd9440aca49be0c620af5
Closes-bug: #1736392