Add driver methods checking for shared storage
The driver interface provides public methods that determine if two
compute nodes use shared storage.

Although we already use a similar mechanism, we do not implement
those public driver methods, for which reason the compute manager
will wrongfully assume that shared storage is never used.

This is not an issue at the moment, as this mechanism is used by
the manager only when cleaning up evacuated instances (which we
do not currently support) and when reverting migrations (in which
case we already have a copy of the instance files in the 'revert
dir').

This will become an issue though if we intend to start cleaning up
migration files when destroying instances.

This change adds the above mentioned public driver methods.

Related-Bug: #1724240

Change-Id: I790b39672fd8ffe467fc9d93daf7b38b26103d5d