Fix assisted volume snapshots race condition
While performing assisted volume snapshots, we're setting the
instance task state to 'image_snapshot_pending' in order to prevent
instance actions that would impact this operation.

The issue is that we're clearing the task state only after calling
back the Cinder API, telling it that the snapshot was created or
deleted. For this reason, a subsequent assisted snapshot request
will be rejected by the Nova API if the task state is not cleared
in time.

This race condition affects quite a few tempest tests that cover
this scenario.

This change ensures that the instance task state is cleared before
calling back the Cinder API.

Closes-Bug: #1739423

Change-Id: I1ae57c109ed551ba03d49b2ac7c6318b3939526d