Evacuate feature
This patch adds the 'evacuate' feature to the compute-hyperv driver.

It works by removing the 'delete_disk_files' from the 'spawn' method,
preventing it from happening every time a new instance is spawned, and
checking whether or not disk files already exist for the given machine.
If they do, they will be used when evacuating the instance on another
host

Change-Id: Ie3e847b9f285cf47b474cdcc96e289c7d053c8a0