Flush console log file before retrieving content
Hyper-V exposes instance serial console ports through named pipes.
We're consuming those named pipes, logging the output to a file,
retrieving the content when the console output is requested.

The issue is that the log file used by os-win never gets flushed
(unless closed). For this reason, we're returning incomplete output.

This change ensures that we're flushing the log file before attempting
to read its content.

Closes-Bug: #1741494

Change-Id: Ic558b16b317e0b1478cc259327d1336f7589afe3
Depends-On: I4b5eb7579ac0c92e9c3630539b0c4d76f5b98c6d