Prepare for readthedocs
Since compute-hyperv is not an official OpenStack project, we cannot
host our docs on docs.openstack.org. For this reason, we'll use
readthedocs.

As readthedocs will not use tox, we have to explicitly provide the doc
build requirements that are not included in requirements.txt.
Unfortunately, we'll have to update this file manually at least once
per release.

Also, reathedocs will invoke sphinx-doc from the /doc dir, for which
reason it cannot find etc/compute-hyperv-config-generator.conf. The
according reference is updated, explicitly declaring the namespaces
to be used when generating the config reference.

Change-Id: Idab6f10f92f7f2e40962f1cabaac8691e6c46619