Fix instance dir check when spawning
When spawning instances, we're checking if the instance dir
already exists, in which case we're logging a message, stating
that existing instance files will be reused.

Well, at least that's what we thought we were doing. Actually,
we're always logging this message as the check is wrong.

This change fixes this issue.

Change-Id: I6cf52ae23d90e687103e8bea0b6c4f9eaba6d36c
Closes-Bug: #1747961