Cluster driver: Blacklist iSCSI/FC volume drivers
There are critical known issues related to the cluster driver and
passthrough disks.

When an instance is failed over, the instance attached volumes need
to be accessible on the destination host. This will not happen
transparently for such disks, for which reason the instance may
bounce between a few nodes and then end up in error state,
basically defeating the purpose of the cluster driver.

This change ensures that the Hyper-V Cluster driver rejects
iSCSI/FC Cinder volumes, raising an error that describes possible
alternatives.

Closes-Bug: #1749958

Change-Id: I231b51c648e1349fcb8df4cd818b2d2887a55d69