Fixes zuul issues
Some issues were not caught during the initial mock autospec
patch which was added to compute_hyperv, and now they're causing
some unit tests to fail. This patch fixes those issues.

Secondly, the docs job relies on tools/tox_install.sh to install
nova, but it is not passing the required arguments, causing it to
fail. This patch updates the script so it can properly execute.

Change-Id: I91cbadb2a05ee4880c3fa761696818ec95c40df2