Enable trusted certificates support
A recent feature has been added, allowing users to specify
trusted certificates (e.g. as barbican secret ids). The idea is
to ensure that the certificates used to sign and validate the
glance image are actually trusted by the user.

In order to enable this feature within our driver, all we have to
do is pass the trusted certificate ids (stored as an instance
object attribute) to the method that fetches glance images.

Blueprint: nova-validate-certificates

Change-Id: Ic28f2b3ecf4ca92dcb7e9643c6e0d207d40b5287