Switch to new instance failover listener
We had a bunch of issues with the WMI based event listener, which
appear to be bugs in the WMI provider.

We've added a new listener that uses the underlying C library. The
only issue is that this new listener cannot tell us the source of
the migrated VMs, so we'll have to rely on the information from
Nova.

Closes-Bug: #1798069
Depends-On: I63d1aa3a6f9fb12d08478eb41fe973b1582b540c

Change-Id: I9e777351eca84274cafc3f04b6bc27c3f6b572ca