Retry plugging ports when clustered instances start
After a while, the Failover Cluster will stop retrying when attempting
to bring back up failed instances. For example, if the CSV is down
more than a few minutes, the cluster groups will be set in "Failed"
state, while the VMs won't be registered on any Hyper-V node.

The issue is that we're only handling cluster group owner changes
(moved instances). If the admin fixes the issue and manually brings
the cluster groups back up, the instances are recreated but we aren't
handling this, so ports won't get reconnected.

This change will double check the ports when clustered instances
start.

Closes-Bug: #1799163

Change-Id: I5caa65d7b7922dc9632b18acedaf1aedeec3fcc3