Move config option definition
The 'move_disks_on_cold_migration' config option is currently
defined in migrationops.py instead of conf.py, for which reason
it's not included in the generated documentation.

This simple change will move it next to the rest of compute-hyperv
config options. While at it, we're adding a help string.

Change-Id: I762156dea8d32bdc13716b88821ed7495e705607