Recreate ports for failed over instances
OVS ports can become invalid after the instances are failed over.
If an instance lands back on the initial host and the ports are
not cleaned up, it will be left with those stale ports.

This change adds a config option, allowing those ports to be recreated.
It will be enabled by default. Deployers may consider disabling it
if the remaining ports are cleaned up when rebooting nodes or if the
network backend is not affected by this issue.

Closes-Bug: #1841778

Change-Id: I58aba53b7bc0a5e8954cf31adac6e7e106944b7e
(cherry picked from commit 7dd16950d3dea87c9853c62fea26437de0dfd814)