Initialize coordination during driver init
At the moment, the "init_host" method initializes the distributed
lock mechanism, yet the driver may try to use distributed locks
before that.

For example, the method that reclaims failed over instances is
called by the driver init method, before init_host. While we
can fix this particular call, it's safer that we initialize the
lock mechanism as early as possible.

Change-Id: If361c0f754d3680726762e3126286227bea6f3f3
Related-Bug: #1841777