Trivial: improve failover logging
We log a warning when detecting a failover for which the destination
host matches the one that's already set in the Nova db. Some times
this means that instance bounced between hosts but most of the times
this is just a migration initiated by Nova, which already updated
the DB.

While at it, when the driver initializes, we'll log the list of
detected failovers before asynchronously handling them.

Change-Id: I808c8cb63ac0011be8d6824a2f8e9292fe39004f
(cherry picked from commit bf39a47dbee4397ea5d183e2c14789de5a107c75)