Handle Python 3 transition and fix tests
While dropping Python 2.7 support, the upstream project has been
updated, so we'll have to move the templates in-tree.

While at it, we're updating the supported Python versions.

Some Nova fixtures have been moved to oslo.test, so in order to be
able to unblock the gate, we'll have to squash those fixes as well.
For the same reason, we'll have to remove nova-network checks, which
has been completely dropped.

A few options have been moved to a separate config group, which
we'll have to take into account.

Change-Id: Ibce6b062a81200611bcaac3f2cb90b14b559375b