Allow limitting concurrent image ops
Nova defines a config option that allows limitting concurrent
image operations. This behavior is disabled by default.

The threshold is enforced using a semaphore, which we're now going
to use.

As opposed to the libvirt driver, we're not applying the semaphore
when merging differencing images since that would have a larger
impact on the hyper-v driver (e.g. when resizing legacy vhd images,
when deleting temporary snapshots, etc). For now, we're using it
only when downloading or uploading glance images.

Change-Id: I74ecf4b343428c71aad77b049faa5d2382e88dc8
blueprint: io-semaphore-for-concurrent-disk-ops