Document image cache manager limitation
When the compute hosts are using different CSVs, Nova must be
configured not to delete unused images since its image caching
mechanism can't properly track the image files in this case.

In particular, instances that were moved to other hosts won't be
tracked and their base images might be removed.

This is not a concern when not using shared storage or when using
a single shared folder (which might as well be a CSV).

Change-Id: If730f59473df576fd32eae5b930c60b1125d7ff1