Drop lower constraints
Maintaining project lower constraints has become an increasingly
difficult task, especially considering the new pip resolver.

That being considered, the TC has stated that it's up to project
maintainers to decide if they want to continue using lower constraints.
The Cinder project has already dropped the lower constraints job [2].

In order to reduce the burden, we dropped indirect dependencies from
the lower constraints file. This is now causing the lower constraints
job to time out:

   This is taking longer than usual. You might need to provide the
   dependency resolver with stricter constraints to reduce runtime.
   See https://pip.pypa.io/warnings/backtracking for guidance.
   If you want to abort this run, press Ctrl + C.

Since we don't really want to go back to maintaining a huge list of
lower constraints, it's time for us to remove the job.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2021-March/021204.html
[2] https://review.opendev.org/c/openstack/cinder/+/789077

Change-Id: I08de76202f9c37332e599b2becd21ca5ca87df19